setup system

+++ setup.js
@@ -1,3 +1,8 @@
 const fs = require('fs')
 
+if (fs.existsSync('index.js')) {
+    console.log('index.js already exists, remove it before setting up !')
+    process.exit(1)
+}
+
 fs.copyFileSync('example.js', 'index.js')
